requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.5
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
phonenumbers==8.12.26
Pillow==8.3.0
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
